Move travel component from common to containers
