+++ src/containers/App/index.js
@@ -7,7 +7,7 @@ import Dashboard from '../Dashboard'
 import About from '../About'
 import Resume from '../Resume'
 // import Interests from '../Interests'
-import Travel from '../Common/Travel'
+import Travel from '../Travel'
 import Contact from '../Contact'
 import Footer from '../Common/Footer'
 import './index.css'
