+++ src/containers/Travel/styles.scss
@@ -1,0 +1,13 @@
+.travel .title {
+  background: #92c14d;
+}
+
+// .travel {
+//   margin-bottom: 40px;
+// }
+
+.travel .travelMap {
+  width: '100%';
+  margin: '0 auto';
+  margin-top: 20px;
+}
